Add files via upload
The source code provided is a multi-threaded consumer-producer program. That means producer threads add items (in this case, a character) to a queue, and consumer threads remove the items from the queue and do something with them (in this case, print the character to standard output). The main thread spawns a consumer thread and a producer thread, then the producer thread also spawns a consumer thread. So in total there are 1 producer thread and 2 consumer threads (plus the main thread). You will find that it compiles correctly, but when you run the program there are in fact several bugs.

For this pre-lab, you will be asked to do two things: First, find and fix the five (5) bugs that have been placed in the source code. (Some of these bugs will appear at runtime, but others may only be apparent by code inspection.) Putting comments near the bugs you found will also help us to grade your pre-lab.

Questions

The main function contains calls to exit() (line 66) and pthread_exit() (l